+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `v0.0.1` Avatar component
+  - Added organization avatar styles
+  - Added `l` size
+
 - `v0.0.10` Form component
   - `v0.0.4` FormEditBanner component
     - `bannerInfo` property to render information in place of a cancel button